simplify the ipi rna seq create etl
